explicit check for bot before adding kspree
Depending on an exception at line 354 wasn't working because id "0" did have a kspree -- it was being set to 0 at the end of checkKSpreeEnd.   It's also being set somewhere else -- sometimes the bots would get on runs and for some reason earn records.  I haven't had the time to figure out under what conditions this happens, so it seems easiest to just check the id before treating the bots as flesh.